feat(include): support @include with multiple parameters
Most blade directives take only one parmeter, but `@include` and it's
kin can take multiple. (In fact, some of these directives aren't valid
without multiple parameters.)

These were not formatted correctly because they were being passed to the
PHP plugin as `(... , ...)` and, because commas are not allowed in that
context, they were being rejected as invalid syntax.

This change works around this by wrapping the content of such directives
in `[]` so that the PHP plugin will validate, parse and format them them
like an array. This looks like it's more-or-less what is happening for
`@forelse` directives already.